updated centinel to a format conducive to package entry
updated centinel to be more like the dev branch by making the
structure more conducive to package entry points
